定时任务增加判断

+++ app/Command/Job.php
@@ -813,12 +813,14 @@ class Job
             }
         }
 
-        # 删除 tg 消息
-        $TelegramTasks = TelegramTasks::where('type', 1)->where('executetime', '<', time())->get();
-        foreach ($TelegramTasks as $Task) {
-            \App\Utils\Telegram\Process::SendPost('deleteMessage', ['chat_id' => $Task->chatid, 'message_id' => $Task->messageid]);
-            TelegramTasks::where('chatid', $Task->chatid)->where('type', '<>', 1)->where('messageid', $Task->messageid)->delete();
-            $Task->delete();
+        if (Config::get('enable_telegram') === true) {
+            # 删除 tg 消息
+            $TelegramTasks = TelegramTasks::where('type', 1)->where('executetime', '<', time())->get();
+            foreach ($TelegramTasks as $Task) {
+                \App\Utils\Telegram\Process::SendPost('deleteMessage', ['chat_id' => $Task->chatid, 'message_id' => $Task->messageid]);
+                TelegramTasks::where('chatid', $Task->chatid)->where('type', '<>', 1)->where('messageid', $Task->messageid)->delete();
+                $Task->delete();
+            }
         }
     }
